Finish caesar_cipher method
The method starts by creating an array of both upper and lower case
letters. Then it splits the string in to an array of characters. Map
then checks whether the character is in either array, and if it is, it
finds the index in whichever array and returns the appropriate letter in
the array given the shift factor. Finally, it joins the string back
together.